Fix read_serial_data signature

+++ serial_port/serial_port_actions.cpp
@@ -811,7 +811,7 @@ bool SerialPortActions::reset_connection()
     return r;
 }
 
-QByteArray SerialPortActions::read_serial_data(uint32_t datalen, unsigned long timeout)
+QByteArray SerialPortActions::read_serial_data(uint32_t datalen, uint16_t timeout)
 {
     if (isDirectConnection())
         return serial_direct->read_serial_data(datalen, timeout);
