+++ serial_port/serial_port_actions.h
@@ -115,7 +115,7 @@ public:
 
     bool reset_connection(void);
 
-    QByteArray read_serial_data(uint32_t datalen, unsigned long timeout);
+    QByteArray read_serial_data(uint32_t datalen, uint16_t timeout);
     QByteArray write_serial_data(QByteArray output);
     QByteArray write_serial_data_echo_check(QByteArray output);
