+++ serial_port/serial_port_actions_direct.h
@@ -96,7 +96,7 @@ public:
 
     void reset_connection();
 
-    QByteArray read_serial_data(uint32_t datalen, unsigned long timeout);
+    QByteArray read_serial_data(uint32_t datalen, uint16_t timeout);
     QByteArray write_serial_data(QByteArray output);
     QByteArray write_serial_data_echo_check(QByteArray output);
